refactor: HostParticipationInfo 컴포넌트 로직 hook으로 분리
