+++ frontend/src/components/participation/HostParticipationInfo.tsx
@@ -1,11 +1,9 @@
-import { useLocation } from 'react-router-dom';
 import { css, styled } from 'styled-components';
 
 import Button from '@Components/common/Button/Button';
 import Input from '@Components/common/Input/Input';
 import Typography from '@Components/common/Typography/Typography';
 
-import useCopyClipBoard from '@Hooks/useCopyClipBoard';
 import useHostParticipationInfo from '@Hooks/useHostParticipationInfo';
 
 import color from '@Styles/color';
@@ -15,20 +13,15 @@ import { ERROR_MESSAGE } from '@Constants/errorMessage';
 import ClipBoard from '@Assets/icons/ClipBoard';
 
 const HostParticipationInfo = () => {
-  const location = useLocation();
-  const studyName = (location.state as { participantCode: string; studyName: string }).studyName;
-  const participantCode = (location.state as { participantCode: string; studyName: string }).participantCode;
-
-  const { nickName, isInputError, handleOnChangeInput } = useHostParticipationInfo();
-  const { onCopy } = useCopyClipBoard();
-
-  const handleOnClickClipBoardButton = () => {
-    onCopy(participantCode).then(() => alert('클립보드에 복사되었습니다.'));
-  };
-
-  const handleOnClickHelperMessage = () => {
-    alert('참여 코드를 스터디원들과 공유하면 스터디원들의 회고와 기록을 확인할 수 있어요.');
-  };
+  const {
+    nickName,
+    isInputError,
+    handleOnChangeInput,
+    participantCode,
+    studyName,
+    handleOnClickClipBoardButton,
+    handleOnClickHelperMessage,
+  } = useHostParticipationInfo();
 
   return (
     <Layout>
